+++ pkg/manager/compressor/service/compress.go
@@ -31,10 +31,10 @@ import (
 type Compressor interface {
 	PreStart(ctx context.Context) error
 	Start(ctx context.Context) <-chan error
-	Compress(ctx context.Context, vector []float64) ([]byte, error)
-	Decompress(ctx context.Context, bytes []byte) ([]float64, error)
-	MultiCompress(ctx context.Context, vectors [][]float64) ([][]byte, error)
-	MultiDecompress(ctx context.Context, bytess [][]byte) ([][]float64, error)
+	Compress(ctx context.Context, vector []float32) ([]byte, error)
+	Decompress(ctx context.Context, bytes []byte) ([]float32, error)
+	MultiCompress(ctx context.Context, vectors [][]float32) ([][]byte, error)
+	MultiDecompress(ctx context.Context, bytess [][]byte) ([][]float32, error)
 }
 
 type compressor struct {
@@ -111,7 +111,7 @@ func (c *compressor) Start(ctx context.Context) <-chan error {
 	return ech
 }
 
-func (c *compressor) dispatchCompress(ctx context.Context, vectors ...[]float64) (results [][]byte, errs error) {
+func (c *compressor) dispatchCompress(ctx context.Context, vectors ...[]float32) (results [][]byte, errs error) {
 	results = make([][]byte, len(vectors))
 
 	wg := new(sync.WaitGroup)
@@ -121,7 +121,7 @@ func (c *compressor) dispatchCompress(ctx context.Context, vectors ...[]float64)
 
 		for iter, vector := range vectors {
 			wg.Add(1)
-			c.jobCh <- func(i int, v []float64) func() error {
+			c.jobCh <- func(i int, v []float32) func() error {
 				return func() error {
 					defer wg.Done()
 
@@ -158,8 +158,8 @@ func (c *compressor) dispatchCompress(ctx context.Context, vectors ...[]float64)
 	return results, errs
 }
 
-func (c *compressor) dispatchDecompress(ctx context.Context, bytess ...[]byte) (results [][]float64, errs error) {
-	results = make([][]float64, len(bytess))
+func (c *compressor) dispatchDecompress(ctx context.Context, bytess ...[]byte) (results [][]float32, errs error) {
+	results = make([][]float32, len(bytess))
 
 	wg := new(sync.WaitGroup)
 	wg.Add(1)
@@ -205,7 +205,7 @@ func (c *compressor) dispatchDecompress(ctx context.Context, bytess ...[]byte) (
 	return results, errs
 }
 
-func (c *compressor) Compress(ctx context.Context, vector []float64) ([]byte, error) {
+func (c *compressor) Compress(ctx context.Context, vector []float32) ([]byte, error) {
 	ress, err := c.dispatchCompress(ctx, vector)
 	if len(ress) != 1 {
 		return nil, err
@@ -214,7 +214,7 @@ func (c *compressor) Compress(ctx context.Context, vector []float64) ([]byte, er
 	return ress[0], err
 }
 
-func (c *compressor) Decompress(ctx context.Context, bytes []byte) ([]float64, error) {
+func (c *compressor) Decompress(ctx context.Context, bytes []byte) ([]float32, error) {
 	ress, err := c.dispatchDecompress(ctx, bytes)
 	if len(ress) != 1 {
 		return nil, err
@@ -223,10 +223,10 @@ func (c *compressor) Decompress(ctx context.Context, bytes []byte) ([]float64, e
 	return ress[0], err
 }
 
-func (c *compressor) MultiCompress(ctx context.Context, vectors [][]float64) ([][]byte, error) {
+func (c *compressor) MultiCompress(ctx context.Context, vectors [][]float32) ([][]byte, error) {
 	return c.dispatchCompress(ctx, vectors...)
 }
 
-func (c *compressor) MultiDecompress(ctx context.Context, bytess [][]byte) ([][]float64, error) {
+func (c *compressor) MultiDecompress(ctx context.Context, bytess [][]byte) ([][]float32, error) {
 	return c.dispatchDecompress(ctx, bytess...)
 }
